+++ store/session.ts
@@ -54,6 +54,9 @@ export class Session {
   /** Theme */
   theme: Theme = 'light';
 
+  /** Vim mode */
+  vimMode: boolean = false;
+
   /** Pine expression to be evaluated */
   expression: string = ''; // observable
 
@@ -115,6 +118,7 @@ export class Session {
   constructor(id: string) {
     this.id = `session-${id}`;
     this.theme = getUserPreference(STORAGE_KEYS.THEME, 'light');
+    this.vimMode = getUserPreference(STORAGE_KEYS.VIM_MODE, false);
 
     makeAutoObservable(this);
 
@@ -240,6 +244,11 @@ export class Session {
   public toggleTheme() {
     this.theme = this.theme === 'light' ? 'dark' : 'light';
     setUserPreference(STORAGE_KEYS.THEME, this.theme);
+  }
+
+  public toggleVimMode() {
+    this.vimMode = !this.vimMode;
+    setUserPreference(STORAGE_KEYS.VIM_MODE, this.vimMode);
   }
 
   public selectNextCandidate(offset: number) {