+++ constants.ts
@@ -7,4 +7,4 @@ export const DEFAULT_SIDEBAR_WIDTH = 240;
 export const MIN_SIDEBAR_WIDTH = 200;
 
 /* Pine Server */
-export const RequiredVersion = '0.16.0';
+export const RequiredVersion = '0.20.0';